Move survey_auto_code into project_reach as a library file
